move analysis to analysis file, fix up
