+++ cedar-policy-core/src/validator/entity_manifest.rs
@@ -659,220 +659,3 @@ pub fn compute_entity_manifest(
     #[allow(clippy::unwrap_used)]
     Ok(manifest.to_typed(validator.schema()).unwrap())
 }
-
-/// A static analysis on type-annotated cedar expressions.
-/// Computes the access paths required to evaluate the expression.
-///
-/// This function populates the provided `AccessDag` store with paths
-/// and returns an `WrappedAccessPaths` analysis result.
-/// The [`WrappedAccessPaths`] contains the result's access paths
-/// and any access paths encountered during the analysis.
-fn analyze_expr_access_paths(
-    expr: &Expr<Option<Type>>,
-    store: &mut AccessDag,
-) -> Result<Rc<WrappedAccessPaths>, EntityManifestError> {
-    Ok(match expr.expr_kind() {
-        ExprKind::Slot(slot_id) => {
-            if slot_id.is_principal() {
-                WrappedAccessPaths::from_root(EntityRoot::Var(Var::Principal), store)
-            } else {
-                assert!(slot_id.is_resource());
-                WrappedAccessPaths::from_root(EntityRoot::Var(Var::Resource), store)
-            }
-        }
-
-        ExprKind::Var(var) => WrappedAccessPaths::from_root(EntityRoot::Var(*var), store),
-
-        ExprKind::Lit(Literal::EntityUID(literal)) =>  WrappedAccessPaths::from_root(
-            EntityRoot::Literal((**literal).clone()),
-            store,
-        ),
-
-        ExprKind::Unknown(_) => Err(PartialExpressionError {})?,
-
-        // Non-entity literals need no fields to be loaded
-        ExprKind::Lit(_) => Rc::new(WrappedAccessPaths::default()),
-
-        ExprKind::If {
-            test_expr,
-            then_expr,
-            else_expr,
-        } => {
-            // For if expressions, the test condition is accessed but not part of the result
-            let test_result = analyze_expr_access_paths(test_expr, store)?;
-            let then_result = analyze_expr_access_paths(then_expr, store)?;
-            let else_result = analyze_expr_access_paths(else_expr, store)?;
-
-            then_result
-                .union(else_result)
-                .with_dropped_paths(test_result)
-        }
-
-        ExprKind::And { left, right }
-        | ExprKind::Or { left, right }
-        | ExprKind::BinaryApp {
-            op: BinaryOp::Less | BinaryOp::LessEq | BinaryOp::Add | BinaryOp::Sub | BinaryOp::Mul,
-            arg1: left,
-            arg2: right,
-        } => {
-            // For these operations, both sides are accessed but the result is a primitive
-            analyze_expr_access_paths(left, store)?.union(analyze_expr_access_paths(right, store)?)
-        }
-
-        ExprKind::UnaryApp { op, arg } => {
-            match op {
-                // These unary ops are on primitive types
-                UnaryOp::Not | UnaryOp::Neg => analyze_expr_access_paths(arg, store)?,
-
-                UnaryOp::IsEmpty => {
-                    let mut arg_result = analyze_expr_access_paths(arg, store)?;
-
-                    // PANIC SAFETY: Typechecking succeeded, so type annotations are present
-                    #[allow(clippy::expect_used)]
-                    let ty = arg
-                        .data()
-                        .as_ref()
-                        .expect("Expected annotated types after typechecking");
-
-                    // For isEmpty, we need all fields of the type
-                    arg_result.require_full_type(ty, store)
-                }
-            }
-        }
-
-        ExprKind::BinaryApp {
-            op:
-                op @ (BinaryOp::Eq
-                | BinaryOp::In
-                | BinaryOp::Contains
-                | BinaryOp::ContainsAll
-                | BinaryOp::ContainsAny),
-            arg1,
-            arg2,
-        } => {
-            // First, find the data paths for each argument
-            let mut arg1_result = analyze_expr_access_paths(arg1, store)?;
-            let mut arg2_result = analyze_expr_access_paths(arg2, store)?;
-
-            // PANIC SAFETY: Typechecking succeeded, so type annotations are present
-            #[allow(clippy::expect_used)]
-            let ty1 = arg1
-                .data()
-                .as_ref()
-                .expect("Expected annotated types after typechecking");
-
-            #[allow(clippy::expect_used)]
-            let ty2 = arg2
-                .data()
-                .as_ref()
-                .expect("Expected annotated types after typechecking");
-
-            // For the `in` operator, we need to handle ancestors
-            if matches!(op, BinaryOp::In) {
-                arg1_result =
-                    arg1_result.with_ancestors_required(&arg2_result.resulting_paths, store);
-            }
-
-            // Load all fields using `full_type_required` for equality checks
-            arg1_result.full_type_required(ty1, store);
-            arg2_result.full_type_required(ty2, store);
-
-            // Union the results and empty the resulting paths since the result is a boolean
-            arg1_result.union(arg2_result)
-        }
-
-        ExprKind::BinaryApp {
-            op: BinaryOp::GetTag | BinaryOp::HasTag,
-            arg1: _,
-            arg2: _,
-        } => Err(UnsupportedCedarFeatureError {
-            feature: "entity tags".into(),
-        }
-        .into()),
-
-        ExprKind::ExtensionFunctionApp { fn_name: _, args } => {
-            // Collect paths from all arguments
-            let mut result = WrappedAccessPaths::default();
-
-            for arg in args.iter() {
-                result = result.union(analyze_expr_access_paths(arg, store)?);
-            }
-
-            // Extension functions return primitives, so empty the resulting paths
-            result
-        }
-
-        ExprKind::Like { expr, pattern: _ }
-        | ExprKind::Is {
-            expr,
-            entity_type: _,
-        } => {
-            // These operations return booleans, so get the result and empty the resulting paths
-            analyze_expr_access_paths(expr, store)?
-        }
-
-        ExprKind::Set(contents) => {
-            let mut result = WrappedAccessPaths::default();
-            let mut element_paths = Vec::new();
-
-            // Collect paths from all set elements
-            for expr in &**contents {
-                let element_result = analyze_expr_access_paths(expr, store)?;
-                result = result.union(element_result.clone());
-                element_paths.push(element_result.resulting_paths);
-            }
-
-            // Combine all element paths into a union
-            let mut combined_paths = Rc::new(WrappedAccessPaths::Empty);
-            for path in element_paths {
-                combined_paths = Rc::new(WrappedAccessPaths::Union(combined_paths, path));
-            }
-
-            // Wrap the combined paths in a SetLiteral
-            result.resulting_paths = Rc::new(WrappedAccessPaths::SetLiteral(combined_paths));
-
-            result
-        }
-
-        ExprKind::Record(content) => {
-            let mut result = WrappedAccessPaths::default();
-            let mut record_contents = HashMap::new();
-
-            // Collect paths from all record fields
-            for (key, child_expr) in content.iter() {
-                let field_result = analyze_expr_access_paths(child_expr, store)?;
-                result = result.union(field_result.clone());
-                record_contents.insert(key.clone(), field_result.resulting_paths);
-            }
-
-            // Create a RecordLiteral with all field paths
-            result.resulting_paths = Rc::new(WrappedAccessPaths::RecordLiteral(record_contents));
-
-            result
-        }
-
-        ExprKind::GetAttr { expr, attr } => {
-            // Get the base expression result
-            let base_result = analyze_expr_access_paths(expr, store)?;
-
-            // Apply the attribute access
-            let result = base_result.get_or_has_attr(attr, store);
-            result
-        }
-
-        ExprKind::HasAttr { expr, attr } => {
-            // Similar to GetAttr, but the result is a boolean
-            let base_result = analyze_expr_access_paths(expr, store)?;
-            let result = base_result.get_or_has_attr(attr, store);
-
-            result
-        }
-
-        #[cfg(feature = "tolerant-ast")]
-        ExprKind::Error { .. } => Err(EntityManifestError::UnsupportedCedarFeature(
-            UnsupportedCedarFeatureError {
-                feature: "No support for AST error nodes".into(),
-            },
-        )),
-    })
-}
