+++ cedar-policy-core/src/validator/entity_manifest/analysis.rs
@@ -2,10 +2,26 @@ use std::{collections::HashMap, rc::Rc};
 
 use smol_str::SmolStr;
 
-use crate::validator::{
-    entity_manifest::{AccessDag, AccessPath, AccessPathVariant, AccessPaths, EntityRoot},
-    types::{EntityRecordKind, Type},
+use crate::{
+    ast::Expr,
+    validator::{
+        entity_manifest::{
+            AccessDag, AccessPath, AccessPathVariant, AccessPaths, EntityManifestError, EntityRoot,
+            PartialExpressionError, UnsupportedCedarFeatureError,
+        },
+        types::{EntityRecordKind, Type},
+    },
 };
+
+use crate::ast::{
+    self, BinaryOp, EntityUID, ExprKind, Literal, PolicySet, RequestType, UnaryOp, Var,
+};
+use crate::entities::err::EntitiesError;
+use miette::Diagnostic;
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+
+use thiserror::Error;
 
 /// Represents [`AccessPath`]s possibly
 /// wrapped in record or set literals.
@@ -23,6 +39,7 @@ pub(crate) enum WrappedAccessPaths {
     /// The union of two [`WrappedAccessPaths`], denoting that
     /// all access paths from both are required.
     /// This is useful for join points in the analysis (`if`, set literals, etc.)
+    /// TODO change Rc to box now that we don't need multiple references to same one
     Union(Rc<WrappedAccessPaths>, Rc<WrappedAccessPaths>),
     /// A record literal, each field having access paths.
     RecordLiteral(HashMap<SmolStr, Rc<WrappedAccessPaths>>),
@@ -121,7 +138,7 @@ impl WrappedAccessPaths {
     /// Union this analysis result with another, taking the union of the resulting paths.
     /// Takes ownership of self and returns self after mutating it.
     pub(crate) fn union(self: Rc<Self>, other: Rc<Self>) -> Rc<Self> {
-        Rc::new(WrappedAccessPaths::Union(Rc::clone(self), Rc::clone(other)))
+        Rc::new(WrappedAccessPaths::Union(self, other))
     }
 
     /// Get all access paths from this wrapped access paths,
@@ -159,6 +176,45 @@ impl WrappedAccessPaths {
                 paths.add_to_access_paths(add_to, include_dropped)
             }
         }
+    }
+
+    /// Get or has tag access paths.
+    /// We can safely assume that self is entity typed.
+    pub(crate) fn get_or_has_tag(
+        self: Rc<Self>,
+        tag_paths: Rc<Self>,
+        store: &mut AccessDag,
+    ) -> Rc<Self> {
+        // compute cross product of the access paths and the tag paths
+        let of_access_paths = self.returned_access_paths().expect(
+            "Tag access paths should not be record or set literals, typechecker should prevent this",
+        );
+        let tag_access_paths = tag_paths
+            .returned_access_paths()
+            .expect("Tag access paths should not be record or set literals, typechecker should prevent this");
+        let mut access_paths = vec![];
+        // cross product of the access paths
+        for of_path in of_access_paths.paths() {
+            for tag_path in tag_access_paths.paths() {
+                // Create a new tag access path
+                let tag_variant = AccessPathVariant::Tag {
+                    of: of_path.clone(),
+                    tag: tag_path.clone(),
+                };
+                // Add the new path to the store
+                let new_path = store.add_path(tag_variant);
+                // Add the new path to the access paths
+                access_paths.push(new_path);
+            }
+        }
+        // now compute the union of all these paths
+        let mut res = Rc::new(WrappedAccessPaths::Empty);
+        // Add the new access paths to the result
+        for path in access_paths {
+            res = res.union(Rc::new(WrappedAccessPaths::AccessPath(path)));
+        }
+        // don't forget to drop self and tag paths, since they represent more paths than just returned access paths
+        res.with_dropped_paths(self).with_dropped_paths(tag_paths)
     }
 
     /// Add accessing this attribute to all access paths
@@ -323,4 +379,202 @@ fn entity_or_record_to_access_paths(
             WrappedAccessPaths::Empty.into()
         }
     }
+}
+
+/// A static analysis on type-annotated cedar expressions.
+/// Computes the access paths required to evaluate the expression.
+///
+/// This function populates the provided `AccessDag` store with paths
+/// and returns an `WrappedAccessPaths` analysis result.
+/// The [`WrappedAccessPaths`] contains the result's access paths
+/// and any access paths encountered during the analysis.
+pub(crate) fn analyze_expr_access_paths(
+    expr: &Expr<Option<Type>>,
+    store: &mut AccessDag,
+) -> Result<Rc<WrappedAccessPaths>, EntityManifestError> {
+    Ok(match expr.expr_kind() {
+        ExprKind::Slot(slot_id) => {
+            if slot_id.is_principal() {
+                WrappedAccessPaths::from_root(EntityRoot::Var(Var::Principal), store)
+            } else {
+                assert!(slot_id.is_resource());
+                WrappedAccessPaths::from_root(EntityRoot::Var(Var::Resource), store)
+            }
+        }
+
+        ExprKind::Var(var) => WrappedAccessPaths::from_root(EntityRoot::Var(*var), store),
+
+        ExprKind::Lit(Literal::EntityUID(literal)) => {
+            WrappedAccessPaths::from_root(EntityRoot::Literal((**literal).clone()), store)
+        }
+
+        ExprKind::Unknown(_) => Err(PartialExpressionError {})?,
+
+        // Non-entity literals need no fields to be loaded
+        ExprKind::Lit(_) => Rc::new(WrappedAccessPaths::default()),
+
+        ExprKind::If {
+            test_expr,
+            then_expr,
+            else_expr,
+        } => {
+            // For if expressions, the test condition is accessed but not part of the result
+            let test_result = analyze_expr_access_paths(test_expr, store)?;
+            let then_result = analyze_expr_access_paths(then_expr, store)?;
+            let else_result = analyze_expr_access_paths(else_expr, store)?;
+
+            then_result
+                .union(else_result)
+                .with_dropped_paths(test_result)
+        }
+
+        ExprKind::And { left, right }
+        | ExprKind::Or { left, right }
+        | ExprKind::BinaryApp {
+            op: BinaryOp::Less | BinaryOp::LessEq | BinaryOp::Add | BinaryOp::Sub | BinaryOp::Mul,
+            arg1: left,
+            arg2: right,
+        } => {
+            // For these operations, both sides are accessed but the result is a primitive
+            analyze_expr_access_paths(left, store)?.union(analyze_expr_access_paths(right, store)?)
+        }
+
+        ExprKind::UnaryApp { op, arg } => {
+            match op {
+                // These unary ops are on primitive types
+                UnaryOp::Not | UnaryOp::Neg => analyze_expr_access_paths(arg, store)?,
+
+                UnaryOp::IsEmpty => {
+                    let mut arg_result = analyze_expr_access_paths(arg, store)?;
+
+                    // PANIC SAFETY: Typechecking succeeded, so type annotations are present
+                    #[allow(clippy::expect_used)]
+                    let ty = arg
+                        .data()
+                        .as_ref()
+                        .expect("Expected annotated types after typechecking");
+
+                    // For isEmpty, we need all fields of the type
+                    arg_result.require_full_type(ty, store)
+                }
+            }
+        }
+
+        ExprKind::BinaryApp {
+            op:
+                op @ (BinaryOp::Eq
+                | BinaryOp::In
+                | BinaryOp::Contains
+                | BinaryOp::ContainsAll
+                | BinaryOp::ContainsAny),
+            arg1,
+            arg2,
+        } => {
+            // First, find the data paths for each argument
+            let mut arg1_result = analyze_expr_access_paths(arg1, store)?;
+            let mut arg2_result = analyze_expr_access_paths(arg2, store)?;
+
+            // PANIC SAFETY: Typechecking succeeded, so type annotations are present
+            #[allow(clippy::expect_used)]
+            let ty1 = arg1
+                .data()
+                .as_ref()
+                .expect("Expected annotated types after typechecking");
+
+            #[allow(clippy::expect_used)]
+            let ty2 = arg2
+                .data()
+                .as_ref()
+                .expect("Expected annotated types after typechecking");
+
+            // For the `in` operator, we need to handle ancestors
+            if matches!(op, BinaryOp::In) {
+                arg1_result = arg1_result.with_ancestors_required(&arg2_result, store);
+            }
+
+            arg1_result
+                .with_dropped_paths(arg2_result.require_full_type(ty2, store))
+                .require_full_type(ty1, store)
+        }
+
+        ExprKind::BinaryApp {
+            op: BinaryOp::GetTag | BinaryOp::HasTag,
+            arg1,
+            arg2,
+        } => {
+            let arg1_result = analyze_expr_access_paths(arg1, store)?;
+            let arg2_result = analyze_expr_access_paths(arg2, store)?;
+
+            arg1_result.get_or_has_tag(arg2_result, store)
+        }
+
+        ExprKind::ExtensionFunctionApp { fn_name: _, args } => {
+            // Collect paths from all arguments
+            let mut result = Rc::new(WrappedAccessPaths::default());
+
+            for arg in args.iter() {
+                result = result.union(analyze_expr_access_paths(arg, store)?);
+            }
+
+            result
+        }
+
+        ExprKind::Like { expr, pattern: _ }
+        | ExprKind::Is {
+            expr,
+            entity_type: _,
+        } => analyze_expr_access_paths(expr, store)?,
+
+        ExprKind::Set(contents) => {
+            let mut combined_paths = Rc::new(WrappedAccessPaths::default());
+
+            // Collect paths from all set elements
+            for expr in &**contents {
+                let element_result = analyze_expr_access_paths(expr, store)?;
+                combined_paths = combined_paths.union(element_result.clone());
+            }
+
+            // Wrap the combined paths in a SetLiteral
+            Rc::new(WrappedAccessPaths::SetLiteral(combined_paths))
+        }
+
+        ExprKind::Record(content) => {
+            let mut result = Rc::new(WrappedAccessPaths::default());
+            let mut record_contents = HashMap::new();
+
+            // Collect paths from all record fields
+            for (key, child_expr) in content.iter() {
+                let field_result = analyze_expr_access_paths(child_expr, store)?;
+                record_contents.insert(key.clone(), field_result);
+            }
+
+            Rc::new(WrappedAccessPaths::RecordLiteral(record_contents))
+        }
+
+        ExprKind::GetAttr { expr, attr } => {
+            // Get the base expression result
+            let base_result = analyze_expr_access_paths(expr, store)?;
+
+            // Apply the attribute access
+            let result = base_result.get_or_has_attr(attr, store);
+            result
+        }
+
+        ExprKind::HasAttr { expr, attr } => {
+            // Similar to GetAttr, but the result is a boolean
+            let base_result = analyze_expr_access_paths(expr, store)?;
+            let result = base_result.get_or_has_attr(attr, store);
+
+            result
+        }
+
+        #[cfg(feature = "tolerant-ast")]
+        ExprKind::Error { .. } => {
+            return Err(EntityManifestError::UnsupportedCedarFeature(
+                UnsupportedCedarFeatureError {
+                    feature: "No support for AST error nodes".into(),
+                },
+            ))
+        }
+    })
 }
